2.Streamlit/assigment1/activity_class/requirements.txt
streamlit
opencv-python
pillow
numpy


